Edded manager
