+++ .gitignore
@@ -93,3 +93,4 @@ iOSInjectionProject/
 
 Pods
 WeatherProject
+WeatherProject
